pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "volkan-project"  # Please adjust the project name if needed
version = "0.1.0"       # Please adjust the project version if needed
requires-python = ">=3.11" # Adjust if using a different Python version
dependencies = [
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "filelock==3.18.0",
    "fonttools==4.57.0",
    "fsspec==2025.3.2",
    "jinja2==3.1.6",
    "joblib==1.4.2",
    "kiwisolver==1.4.8",
    "markupsafe==3.0.2",
    "matplotlib==3.10.1",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "numpy==2.2.4",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "pykan==0.2.8",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "six==1.17.0",
    "sympy==1.13.1",
    "threadpoolctl==3.6.0",
    "torch==2.6.0",
    "tqdm==4.67.1",
    "triton==3.2.0",
    "typing-extensions==4.13.1",
    "tzdata==2025.2",
]

[project.urls]
"Homepage" = "https://github.com/example/repository" # Optional: Add your project homepage URL 